unit:(tests) mock home dir as temp dir
Closes https://github.com/chef/chef-analyze/issues/57
